# backend/docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/architecture/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/OpenByteInc/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/strategies/script-templates:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: getApiStrategiesScriptTemplates
      summary: Get Api Strategies Script Templates
      x-visibility: internal
  /api/strategies/script-sources:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: getApiStrategiesScriptSources
      summary: Get Api Strategies Script Sources
      x-visibility: internal
  /api/strategies/script-sources/detail:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: getApiStrategiesScriptSourcesDetail
      summary: Get Api Strategies Script Sources Detail
      x-visibility: internal
  /api/strategies/script-sources/create:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesCreate
      summary: Post Api Strategies Script Sources Create
      x-visibility: internal
  /api/strategies/script-sources/update:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesUpdate
      summary: Post Api Strategies Script Sources Update
      x-visibility: internal
    put:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: putApiStrategiesScriptSourcesUpdate
      summary: Put Api Strategies Script Sources Update
      x-visibility: internal
  /api/strategies/script-sources/compile:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesCompile
      summary: Post Api Strategies Script Sources Compile
      x-visibility: internal
  /api/strategies/script-sources/delete:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesDelete
      summary: Post Api Strategies Script Sources Delete
      x-visibility: internal
    delete:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: deleteApiStrategiesScriptSourcesDelete
      summary: Delete Api Strategies Script Sources Delete
      x-visibility: internal
  /api/strategies/script-sources/versions:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: getApiStrategiesScriptSourcesVersions
      summary: Get Api Strategies Script Sources Versions
      x-visibility: internal
  /api/strategies/script-sources/versions/{version_id}:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: getApiStrategiesScriptSourcesVersionsVersionId
      summary: Get Api Strategies Script Sources Versions Version Id
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: internal
    parameters:
    - in: path
      name: version_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/strategies/script-sources/versions/restore:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesVersionsRestore
      summary: Post Api Strategies Script Sources Versions Restore
      x-visibility: internal
  /api/strategies/script-sources/publish:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Strategy
      operationId: postApiStrategiesScriptSourcesPublish
      summary: Post Api Strategies Script Sources Publish
      x-visibility: internal
...
openapi: 3.0.3